requirements.txt
# for maintaining requirements.txt; `pipreqs --encoding=utf8 --print`
pipreqs==0.4.13

streamlit==1.31.0
python-dotenv==1.0.0

pandas==2.2.0 # used for Streamlit tables
plotly==5.18.0  # used for Streamlit plots
numpy==1.26.3  # used for chart_utils